flatten arch header and code
