+++ libosal/include/libosal.h
@@ -18,6 +18,7 @@
 #include "osal_storage_buffer.h"
 #include "osal_storage_vfs.h"
 #include "osal_storage_driver_file.h"
+#include "osal_storage_driver_asset.h"
 #include "osal_ui_service.h"
 #include "osal_ui_widget.h"
 #include "osal_ui_pane.h"
