+++ libosal/include/osal_util_jni.h
@@ -1,5 +1,6 @@
-#ifndef __OSAL_PLATFORM_ANDROID_JNI_H__
-#define __OSAL_PLATFORM_ANDROID_JNI_H__
+#ifndef __OSAL_UTIL_JNI_H__
+#define __OSAL_UTIL_JNI_H__
+#if defined (PLATFORM_ANDROID)
 
 #include <jni.h>
 
@@ -49,4 +50,5 @@ private:
 // ----------------------------------------------------------------------------
 } // namespace osal
 
+#endif // PLATFORM_ANDROID
 #endif // __OSAL_PLATFORM_ANDROID_JNI_H__
